Extracted isOpen function to separate file and refactored it
