+++ .gitignore
@@ -17,6 +17,7 @@
 
 # misc
 .DS_Store
+.vscode
 
 # debug
 npm-debug.log*
